Refactoring: get territory map from JSON file
Territories now have readable IDs - drawn on the map image itself and as keys in the JSON file. All info on the territories - including color, neighbors and containing continents - is parsed from the JSON file, in `TerritoryMap`.

`Continent` has been updated to store a list of the territories it contains, and `Territory` has been updated to store a list of its neighbors.

Partly transparent pixels in the map image are now made either completely transparent or opaque on runtime, to allow writing "comments" in the image.

Also, added Jackson to `core` dependencies.

#42