Made views not recreate (redundant) data
Every view that implemented of `AbstractView` used to recreate data that did not need to be recreated (like fonts) and data that the view was never going to use (like fonts specifically for other views) through the `create()` method. To fix this, renamed `AbstractView` to `UIStyle` and made it a singleton that every view can access and that is initialized and disposed by `RiskyRisk`.

Also removed a couple unused methods in `UIStyle`.

Still have to fix some views modifying `UIStyle`'s public data, like `PhaseView` does in `onNextPlayer()`.