MultiplayerModel contains territories per player
Made the "leaderboard"-map of `MultiplayerModel` contain sets of the territories players own, instead of simply the number of territories they own. This facilitates easier querying both which territories a player owns, and the number of them.

Also implemented sorting of the leaderboard, so that the players with the most territories get placed at the top.